% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.R
\name{GSDF}
\alias{GSDF}
\title{Gene Specific data filter}
\usage{
GSDF(expression, methylation, miRNA, target_gene)
}
\arguments{
\item{expression}{expression datframe}

\item{methylation}{methykation dataframe}

\item{miRNA}{miRNA dataframe}

\item{target_gene}{gene symbol}
}
\value{
GSD
}
\description{
Gene Specific data filter
}
