% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMImR_import.R
\name{import_EMImR}
\alias{import_EMImR}
\title{Import files for the EMImR function}
\usage{
import_EMImR(path)
}
\arguments{
\item{path}{file that conatins data csv files}
}
\value{
expression, methylation and miRNA
}
\description{
Import files for the EMImR function
}
