% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF_Import.R
\name{import_GSDF}
\alias{import_GSDF}
\title{import files for GSDF function}
\usage{
import_GSDF(path)
}
\arguments{
\item{path}{file that contains the data csv files}
}
\value{
expression, methylation,  miRNA, DEGs , DMGs and GDEIRs
}
\description{
import files for GSDF function
}
